Add a first prototype of Nested Input in OperationOperator with ClosureOp
We have now a new operation: ClosureOp. With this operation it's possible
to compute and pass more than one OMEMOBJ in the NestedOp.
While NestedOp give to the flow a sort of "height" with the execution of
more than one Op, the ClosureOp give the "weight" of execute more than one
op at the same time in a NestedOp flux.