Refactor StackObject, RegistryColl, getStackPosition, getRegPosition
With this edit, one of the last of the Environment Update, the
containers StackObject and RegistryColl are been refactored in protected
members with relative property plus StackObject's name changed to
"StackColl".
The getStackPosition and getRegPosition haven't now more the ability to
take as argument list of MemObj, but only a single one. The work of
applying to multiple instances have been delegated to the caller.